+++ app/components/Carousel/ControlsContainer/index.js
@@ -1,0 +1,1 @@
+export { default as ControlsContainer } from './ControlsContainer';
